+++ examples/remote_twincat.cpp
@@ -3,7 +3,7 @@
 // Created Date: 12/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 26/09/2022
+// Last Modified: 07/11/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -24,7 +24,7 @@ int main() try {
 
   autd.open(std::move(link));
 
-  return run(std::move(autd));
+  return run(autd);
 } catch (std::exception& e) {
   std::cerr << e.what() << std::endl;
   return -1;