+++ rust/autd3-core/src/geometry/builder.rs
@@ -4,7 +4,7 @@
  * Created Date: 27/04/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 01/06/2022
+ * Last Modified: 07/11/2022
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -41,7 +41,7 @@ impl GeometryBuilder<Normal> {
     pub fn new() -> Self {
         Self {
             attenuation: 0.0,
-            sound_speed: 340.0,
+            sound_speed: 340.0e3,
             _mode: PhantomData,
         }
     }