+++ rust/autd3/src/prelude.rs
@@ -4,7 +4,7 @@
  * Created Date: 27/04/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 01/06/2022
+ * Last Modified: 07/11/2022
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -14,10 +14,11 @@
 pub use crate::{controller::Controller, gain::*, modulation::*};
 
 pub use autd3_core::{
+    amplitude::Amplitudes,
     delay::ModDelay,
     geometry::{
-        Amplitudes, Geometry, GeometryBuilder, LegacyTransducer, NormalPhaseTransducer,
-        NormalTransducer, Transducer, Vector3,
+        Geometry, GeometryBuilder, LegacyTransducer, NormalPhaseTransducer, NormalTransducer,
+        Transducer, Vector3,
     },
     link::Link,
     silencer_config::SilencerConfig,